LPS-30949 Avoiding duplication of first instance when editing time of recurring event
If one edits a recurring event by clicking on its first instance and then choose
to salve "All Following", the first instance would be duplicated with the previous
time. This commit solves it.

Also, now the loop does not update the startTimeJCalendar anymore, so multiple, undesired
changes are are not applied to it anymore. It used to result on events from child
calendars having one or more day less than the edited master calendar.